FPP GPIO Toggle Initital commit
Put simply this plugin allows you to toggle regular GPIO Channel Outputs

I created this because I had a NC relay hooked 240v Contactor and needed
a way to Open and close it so I can toggle power on that output. I run a
remote show (20mins drive) and had a issue in 2016 where I sometimes
needed to power cycle an output on numerous occasions.

Toggling is only available on RPi since I use wiringPi's gpio read to
find out the current state of an output when toggling.